Only show and allow modification when applicable message is supported
When message is unsupported the controls are hidden and settings cannot  be modified.
Before this modification, the returned `unsupported error` from the FC was discarded and an infinite retry to get the data was the result.
This is to cover complete absence of a message, API version checks are still required when the content of a supported message can vary.