Issue #55: prop balance UI
For each motor add a control switch and the RMS reading.
The switch controls both the motor and the refreshing of the RMS
reading.
The aim is to turn all switches off, predefine a throttle value, turn on
each switch one by one for about 10s each. The result is the RMS value
of each motor.