Delete backup/restore functionality.
Reasons:

1) it's broken.
2) it doesn't back up everything.
3) the migration code is a nightmare to maintain and has been largely
ignored by developers when they add new features.
4) it may prompt some developers to create a better solution.  e.g. use
the parameter groups functionality in CF 1.13.0 and newer for backups.
parameter groups were specifically designed to be backed up and restored
by the FC and importantly, have version numbers in them.

Feel free to reference the code in this commit that was deleted for any
new system.