CF - Use release `name` field for version.
the `version` field relates to the git TAG, the `name` field is
user-supplied when creating a release.

This allows us to create better tag names as there will be overlap in
the future unless the tag naming scheme is changed.

Tag names for cleanflight are now 'CLFL_vx.x.x'.  release names are as
before 'vx.x.x'